modificando os campos para fazer  as chamadas no ClienteController.kt

+++ src/main/kotlin/com/example/SistemaDeVendasDeAutomoveis/Controller/ClienteController.kt
@@ -2,6 +2,8 @@ package com.example.SistemaDeVendasDeAutomoveis.Controller
 
 import com.example.SistemaDeVendasDeAutomoveis.Domain.ClienteDomain
 import com.example.SistemaDeVendasDeAutomoveis.Repository.ClienteRepository
+import com.example.SistemaDeVendasDeAutomoveis.Request.ClienteRequest
+import com.example.SistemaDeVendasDeAutomoveis.Response.ClientResponse
 
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
@@ -18,13 +20,13 @@ class ClienteController(val repository: ClienteRepository) {
     fun search() : List<ClienteDomain> = repository.findAll()
 
     @GetMapping("/{id}")
-    fun searchById(@PathVariable idCliente: Long) : ResponseEntity<ClienteDomain> =
+    fun searchById(@PathVariable idCliente: Int) : ResponseEntity<ClienteDomain> =
         repository.findByidCliente(idCliente).map {
             ResponseEntity.ok(it)
         }.orElse(ResponseEntity.notFound().build())
 
     @PutMapping("/{id}")
-    fun upadate(@PathVariable idCliente: Long, @RequestBody clienteDomain: ClienteDomain) : ResponseEntity<ClienteDomain> =
+    fun upadate(@PathVariable idCliente: Int, @RequestBody clienteDomain: ClienteDomain) : ResponseEntity<ClienteDomain> =
         repository.findByidCliente(idCliente).map {
             val clienteUpdate = it.copy(
                 nome = clienteDomain.nome,
@@ -36,7 +38,7 @@ class ClienteController(val repository: ClienteRepository) {
         }.orElse(ResponseEntity.notFound().build())
 
     @DeleteMapping("/{id}")
-    fun delete(@PathVariable idCliente: Long) : ResponseEntity<Void> =
+    fun delete(@PathVariable idCliente: Int) : ResponseEntity<Void> =
         repository.findByidCliente(idCliente).map {
             repository.delete(it)
             ResponseEntity<Void>(HttpStatus.OK)
