+++ src/main/kotlin/com/example/SistemaDeVendasDeAutomoveis/Repository/ClienteRepository.kt
@@ -4,7 +4,7 @@ import com.example.SistemaDeVendasDeAutomoveis.Domain.ClienteDomain
 import org.springframework.data.jpa.repository.JpaRepository
 import java.util.*
 
-interface ClienteRepository : JpaRepository<ClienteDomain, Long>{
+interface ClienteRepository : JpaRepository<ClienteDomain, Int>{
 
-    fun findByidCliente (idCliente: Long): Optional<ClienteDomain>
+    fun findByidCliente (idCliente: Int): Optional<ClienteDomain>
 }
